backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
firebase-admin
psycopg2-binary
pytest
pytest-cov
python-dotenv
black
bandit
